Fix overloaded funtion to call I2C master begin

+++ src/Module_4RELAY.cpp
@@ -7,7 +7,7 @@ bool MODULE_4RELAY::begin(TwoWire* wire, uint8_t addr, uint8_t sda, uint8_t scl,
     _addr = addr;
     _sda  = sda;
     _scl  = scl;
-    _wire->begin(_sda, _scl);
+    _wire->begin((int)_sda, (int)_scl);
     delay(10);
     _wire->beginTransmission(_addr);
     uint8_t error = _wire->endTransmission();
